\name{dataG}
\alias{dataG}
\docType{data}
\title{dataG
}
\description{
Dataset G for the usage examples.
}
\usage{data("dataG")}
\format{
  A data frame with 1000 observations on the following 8 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{Y1}}{a numeric vector}
    \item{\code{Y2}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}
\examples{
data(dataG)
## maybe str(dataG) ; plot(dataG) ...
}
\keyword{datasets}
