\name{dataS}
\alias{dataS}
\docType{data}
\title{dataS
}
\description{
Dataset S for the usage examples.
}
\usage{data("dataS")}
\format{
  A data frame with 1000 observations on the following 4 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{Y1}}{a numeric vector}
    \item{\code{Y2}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}
\examples{
data(dataS)
## maybe str(dataS) ; plot(dataS) ...
}
\keyword{datasets}
