\name{pleiotropy-package}
\alias{pleiotropy-package}
\alias{pleiotropy}
\docType{package}
\title{
\packageTitle{pleiotropy}
}
\description{
\packageDescription{pleiotropy}
}

\author{
\packageAuthor{pleiotropy}

Maintainer: \packageMaintainer{pleiotropy}
}
\references{
Lutz SM, Fingerlin T, John E Hokanson, Lange C. (2016) A General
Approach to Testing for Pleiotropy with Rare and Common Variants.
(Submitted Manuscript).
}

\keyword{ package }